Use class-based API for lazy tests
