+++ lib/graphql/deprecated_dsl.rb
@@ -1,42 +1,26 @@
 # frozen_string_literal: true
 module GraphQL
+  # There are two ways to apply the deprecated `!` DSL to class-style schema definitions:
+  #
+  # 1. Scoped by file (CRuby only), add to the top of the file:
+  #
+  #      using GraphQL::DeprecatedDSL
+  #
+  # 2. Global application, add before schema definition:
+  #
+  #      GraphQL::DeprecatedDSL.activate
+  #
   module DeprecatedDSL
-    refine GraphQL::SchemaMember.singleton_class do
+    def self.activate!
+      GraphQL::SchemaMember.extend(Methods)
+    end
+    module Methods
       def !
         to_non_null_type
       end
-
-      def to_list_type
-        ListTypeProxy.new(self)
-      end
-
-      def to_non_null_type
-        NonNullTypeProxy.new(self)
-      end
-
-      class ListTypeProxy
-        include GraphQL::SchemaMember::CachedGraphQLDefinition
-
-        def initialize(member)
-          @member = member
-        end
-
-        def to_graphql
-          @member.graphql_definition.to_list_type
-        end
-      end
-
-      class NonNullTypeProxy
-        include GraphQL::SchemaMember::CachedGraphQLDefinition
-
-        def initialize(member)
-          @member = member
-        end
-
-        def to_graphql
-          @member.graphql_definition.to_non_null_type
-        end
-      end
+    end
+    refine GraphQL::SchemaMember.singleton_class do
+      include Methods
     end
   end
 end
