+++ lib/graphql/schema_member.rb
@@ -83,6 +83,38 @@ module GraphQL
       def to_graphql
         raise NotImplementedError
       end
+
+      def to_list_type
+        ListTypeProxy.new(self)
+      end
+
+      def to_non_null_type
+        NonNullTypeProxy.new(self)
+      end
+    end
+
+    class ListTypeProxy
+      include GraphQL::SchemaMember::CachedGraphQLDefinition
+
+      def initialize(member)
+        @member = member
+      end
+
+      def to_graphql
+        @member.graphql_definition.to_list_type
+      end
+    end
+
+    class NonNullTypeProxy
+      include GraphQL::SchemaMember::CachedGraphQLDefinition
+
+      def initialize(member)
+        @member = member
+      end
+
+      def to_graphql
+        @member.graphql_definition.to_non_null_type
+      end
     end
   end
 end
